refactor(web): retire /dashboard — move milestones to /milestones, redirect
Per user direction: the analytics-and-milestones Dashboard wasn't
pulling its weight. Each section had a more natural home elsewhere,
so the page is now retired.

What moved where
- Milestones → new /milestones page (this commit). Self-contained:
  owns its data fetch + the streak/stats derivations useAchievements
  needs, so it doesn't drag in the broader Dashboard data hook.
  Hooks use-achievements + use-milestone-celebrations relocated from
  /dashboard/_hooks → /milestones/_hooks (their only consumer).
- Genres / taste analytics → already covered by /wrapped (which now
  also has the monthly variant after dc9e458). No port needed.
- Recent picks → already covered by /history. Scrapped from web.
- In-progress / overview suggestions → already covered by the
  Finish-What-You-Started banner on /feed and the Feed quiz menu.

Route hygiene
- /dashboard now server-redirects to /feed (preserves bookmarks,
  shared deep links, mobile pushes).
- Dashboard internals deleted: _components (dashboard-tab-content,
  pick-card, recommendation-modal) and the two non-milestones hooks
  (use-dashboard-data, use-dashboard-copy). Nothing outside /dashboard
  imported these.
- AppNavbar: replace the "Dashboard" item with "Milestones".
- ROUTES constant: drop DASHBOARD, add FEED + MILESTONES.
- share-target landing: secondary CTA points to /feed.
- Auth fallback redirect: defaults to /feed when no ?next= is given.
- New Navbar.items.milestones + Dashboard.body.milestones.pageTitle /
  backToFeed strings (en + es). The Dashboard i18n namespace itself
  is intentionally left in place — /milestones still reads from
  Dashboard.body.milestones.* (single source of truth, no churn).

Typecheck, lint, 104 tests still green.
