+++ src/lib/constants.ts
@@ -5,7 +5,8 @@ export const ROUTES = {
   AUTH_CALLBACK: "/auth/callback",
   AUTH_RESET_PASSWORD: "/auth/reset-password",
   AUTH_VERIFIED: "/auth/verified",
-  DASHBOARD: "/dashboard",
+  FEED: "/feed",
+  MILESTONES: "/milestones",
   DEMO: "/demo",
   DEMO_RESULTS: "/demo/results",
   QUIZ: "/quiz",
